+++ ft_striteri.c
@@ -6,16 +6,15 @@
 /*   By: yooh <yooh@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/10 15:43:03 by yooh              #+#    #+#             */
-/*   Updated: 2022/11/10 15:44:51 by yooh             ###   ########.fr       */
+/*   Updated: 2022/11/14 10:39:16 by yooh             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
 #include "libft.h"
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	unsigned int		i;
+	size_t	i;
 
 	if (!s)
 		return ;
