+++ ft_strjoin.c
@@ -6,19 +6,18 @@
 /*   By: yooh <yooh@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 17:50:07 by yooh              #+#    #+#             */
-/*   Updated: 2022/11/10 11:16:19 by yooh             ###   ########.fr       */
+/*   Updated: 2022/11/14 10:39:38 by yooh             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
 #include "libft.h"
 
 char	*ft_strjoin(char const *s1, char const *s2)
 {
 	const size_t	len = ft_strlen(s1) + ft_strlen(s2);
 	char			*result;
-	int				i;
-	int				j;
+	size_t			i;
+	size_t			j;
 
 	i = -1;
 	j = -1;
